% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjust_network.R
\name{adjust_network}
\alias{adjust_network}
\title{adjust_network}
\usage{
adjust_network(fixed, obs, output = NA)
}
\arguments{
\item{fixed}{data.frame; giving the coordinates to the fixed (known) points in the network.}

\item{obs}{data.frame; giving the measured distance and bearing to the known and unknown points.}

\item{output}{optional filename for the output from gama-local.}
}
\value{
a data.frame with the coordinates for all the points in the network, including the known points.
Optionally a file with the output from the adjustment in gama-local (adjustment errors etc.).
}
\description{
run adjustment of a geodetic free network
}
\note{
the input coordinates of the known points should be formated as a three-column data.frame
with three columns

 \itemize{
 \item \bold{id}  name of the point
 \item \bold{x}  x (east) coordinate of the point
 \item \bold{y}  y (north) coordinate of the point
	}

The input distances and bearings should be formated as a data.frame with four colums:

 \itemize{
 \item \bold{from}  name of the point from where the observation is done
 \item \bold{to}  name of the observed point
 \item \bold{direction}  bearing to the observed point
 \item \bold{distance}  distance to the observed point
	}
}
\author{
Marius Hauglin 2016-2017 \email{marius.hauglin@gmail.com}
}
\references{
Aleš Čepek, GNU gama project, https://www.gnu.org/software/gama/
}

