% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1get_sequence_information.R
\name{get_sequence_information}
\alias{get_sequence_information}
\title{Retrieves sequence information for a given list of accessions.}
\usage{
get_sequence_information(accession, remove_dot_1 = FALSE)
}
\arguments{
\item{accession}{accession number.}

\item{remove_dot_1}{Logical. If TRUE, removes '.1' from accession numbers.}
}
\value{
A data frame containing sequence information for the given accessions.
}
\description{
Retrieves sequence information for a given list of accessions.
}
\examples{
\dontrun{
accessions_to_query <- sam_acc$accession
seq_info <- get_sequence_information(accessions_to_query, remove_dot_1 = TRUE)
print(seq_info)
}
}
