% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Writes a data frame to a FASTA file.}
\usage{
write_fasta(data)
}
\arguments{
\item{data}{A data frame with 'SequenceID' and 'Sequence' columns.}
}
\value{
A character vector representing the contents of the FASTA file.
}
\description{
This function takes a data frame with 'SequenceID' and 'Sequence' columns
and writes the contents to a FASTA file. Each row in the data frame corresponds
to a sequence in the FASTA file, with the 'SequenceID' used as the header line
and the 'Sequence' as the sequence line.
}
\examples{
\dontrun{
# Example usage:
data <- final_data

# Call the function
fasta_content <- write_fasta(data)

# Print or use the `fasta_content` as needed
print(fasta_content)

# Specify the output file path
output_file_path <- "output.fasta"

# Open a connection to the output file
output_file <- file(output_file_path, "w")

# Write the content to the file
writeLines(fasta_content, output_file)

# Close the output file
close(output_file)

# Print a message indicating successful file creation
cat("FASTA file has been created and saved at:", output_file_path, "\n")
}
}
