% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3SampleID_vs_NumSequences.R
\name{SampleID_vs_NumSequences}
\alias{SampleID_vs_NumSequences}
\title{Plots the number of sequences and the probability of sequences per SampleID.}
\usage{
SampleID_vs_NumSequences(final_data)
}
\arguments{
\item{final_data}{A data frame with 'SampleID' and 'SequenceID' columns.}
}
\value{
A list containing two ggplot2 plots: 'plot_num_sequences' and 'plot_prob'.
}
\description{
This function takes a data frame with 'SampleID' and 'SequenceID' columns and
creates two bar plots. The first plot shows the number of sequences per SampleID,
and the second plot shows the probability of sequences per SampleID.
}
\examples{
\dontrun{
# Example usage
plots <- SampleID_vs_NumSequences(final_data)

# Set the file name for the TIFF images
tiff_file_num_sequences <- "output_directory/0. SampleID_vs_NumSequences.tiff"
tiff_file_prob <- "output_directory/0. SampleID_vs_Probability.tiff"

# Set the width, height, and DPI parameters
width_inch <- 8
height_inch <- 8
dpi <- 300

# Open the TIFF devices and save the plots
tiff(tiff_file_num_sequences, width = width_inch, height = height_inch, units = "in", res = dpi)
print(plots$plot_num_sequences)
dev.off()

tiff(tiff_file_prob, width = width_inch, height = height_inch, units = "in", res = dpi)
print(plots$plot_prob)
dev.off()
}
}
