% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4data_sampling.R
\name{data_sampling}
\alias{data_sampling}
\title{Samples data from each SampleID group if specified, otherwise uses the final data.}
\usage{
data_sampling(final_data, sample_proportion = NULL)
}
\arguments{
\item{final_data}{A data frame with 'SampleID' and 'SequenceID' columns.}

\item{sample_proportion}{Proportion of data to sample from each SampleID group.
If NULL, the original data frame is returned.}
}
\value{
A data frame either with the original data or sampled data based on the specified proportion.
}
\description{
This function takes a data frame with 'SampleID' and 'SequenceID' columns and
either returns the original data frame (if sample_proportion is NULL) or
samples a specified proportion from each SampleID group.
}
\examples{
\dontrun{
# Example usage:
# If you want to use the final data:
final_data <- preprocess_for_alignment(sam_acc, seq_info)$final_data

# If you want to sample 10\% from each SampleID group:
sampled_data <- data_sampling(final_data, sample_proportion = 0.1)
}
}
