% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_by_distances.R
\name{find_by_distances}
\alias{find_by_distances}
\title{find_by_distances}
\usage{
find_by_distances(x1, y1, x2, y2, dist1, dist2)
}
\arguments{
\item{x1}{numeric; x coordinate of the first known point.}

\item{y1}{numeric; y coordinate of the first known point.}

\item{x2}{numeric; x coordinate of the second known point.}

\item{y2}{numeric; y coordinate of the second known point.}

\item{dist1}{numeric; distance to the first known point.}

\item{dist2}{numeric; distance to the second known point.}
}
\value{
coordinates of the two possible positions
}
\description{
find the two possible positions of one unknown point using the distances to two known points
}
\author{
Marius Hauglin 2016 \email{marius.hauglin@gnmail.com}
}
\references{
http://gis.stackexchange.com/questions/112905/map-location-of-unknown-point-with-distances-to-two-known-points
}

