\name{flame}
\docType{data}
\alias{flame}
\title{DNA microarray data.}
\description{
  All description about the data set may be found in \emph{FLAME, a novel fuzzy clustering method for the analysis of DNA microarray data} article, and the Clustering basic benchmark site, in the references.
}
\usage{data(flame)}
\format{A data frame containing 240 observations and two dimensions, forming two partitions:

    \enumerate{
    \item x1: real positive values
    \item x2: real positive values
    }

   Originally, the dataset had contained three dimensions. We intentionally removed the third dimension that corresponds to the label which the data point belongs.
}

\source{The dataset was collected from \href{http://cs.joensuu.fi/sipu/datasets/}{Clustering basic benchmark} site. }

\references{
L. Fu and E. Medico, \emph{FLAME, a novel fuzzy clustering method for the analysis of DNA microarray data}. BMC bioinformatics, 2007. 8(1): p. 3.

P. Franti and S. Sieranoja, \emph{K-means properties on six clustering benchmark datasets,} vol. 48, no. 12. pp. 4743-4759, 2018.

}
\keyword{datasets}
