% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5alignment_info.R
\name{alignment_info}
\alias{alignment_info}
\title{Sequence Alignment and Analysis}
\usage{
alignment_info(data, type = "global", verbose = 1)
}
\arguments{
\item{data}{A data frame with 'SequenceID' and 'Sequence' columns.}

\item{type}{A character string specifying the type of sequence alignment ('global', 'local', etc.).}

\item{verbose}{An integer indicating the level of verbosity (0, 1, or 2).}
}
\value{
A list containing matrices and results from sequence alignment.
}
\description{
Sequence Alignment and Analysis
}
\examples{
\dontrun{
alignment_results <- alignment_info(final_data, type = "global", verbose = 1)

# Access the resulting data frames
score_matrix <- alignment_results$score_matrix
normalized_score_matrix <- alignment_results$normalized_score_matrix

total_aligned_positions_matrix <- alignment_results$total_aligned_positions_matrix
number_of_matching_positions_matrix <- alignment_results$number_of_matching_positions_matrix

percent_similarity_matrix <- alignment_results$percent_similarity_matrix

alignment_results_list <- alignment_results$alignment_results_list

alignment_info_matrix <- alignment_results$alignment_info_matrix


save_path <- "output_directory/"
# Save the list of alignment results to an RDS file
saveRDS(alignment_results_list, file.path(save_path, "alignment_results_list.rds"))

# Save matrices to files
write.table(score_matrix, file.path(save_path, "score_matrix.txt"), sep = "\t")
}
}
