\name{gama-package}
\alias{gama-package}
\alias{gama}
\docType{package}
\title{
gama - adjustment of a geodetic free network 
}
\description{
Run adjustment of a geodetic free network using the gama-local tool from the GNU gama project.
}
\details{
\tabular{ll}{
Package: \tab gama\cr
Type: \tab Package\cr
Version: \tab 0.7\cr
Date: \tab 2017-06-16\cr
License: \tab What license is it under?\cr
}
}
\author{
Marius Hauglin
Maintainer: Marius Hauglin <marius.hauglin@gmail.com>

}
\references{
Aleš Čepek, GNU gama project, https://www.gnu.org/software/gama/
}
\keyword{ 
geodetic network
geodesy
}

