\name{gama.plot.partitions}
\alias{gama.plot.partitions}
\title{Plots results of a Gama clustering.}
\description{
  This function takes a gama object and plots the partitions found by the algorithm. The partitions will be coloured accordingly the distribution of the clusters.
}
\usage{
gama.plot.partitions(gama.obj = NULL, view = "pca", \dots)
}
\arguments{
  \item{gama.obj}{an object of type 'gama' generated by an appropriate gama call to gama clustering.}
  \item{view}{the representation used to plot the partitions. If 'pca' is used, the function will perform a Principal Component Analysis over the dataset and plot the two main components. If 'total.sum' is used, the function will sum all dimensions for each observation and plot this sum, highlighting its partitions. If omitted, the default value 'pca' will be used.}
  \item{\dots}{other arguments that user may pass to the function.}

}
\references{
Mardia, K. V., J. T. Kent, and J. M. Bibby (1979) Multivariate Analysis, London: Academic Press.
}
\seealso{
  \code{\link{gama}, \link{print.gama}}.
}

\examples{
# loads data about CPU execution metrics of a distributed
# version of Alternating Least Squares (ALS) algorithm
library(gama)
data(cpu.als)

# the gama clustering algorithm call
gamaObj <- gama(cpu.als, k = 3)


# a call to gama.plot.partitions function with "Principal Component Analysis" view method
gama.plot.partitions(gamaObj, view.method = "pca")

# a call to gama.plot.partitions function with "Total Sum" of dimensions view method
\dontrun{gama.plot.partitions(gamaObj, view.method = "total.sum")}
}
\keyword{file}
