% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6compute_average_similarity_matrix.R
\name{compute_average_similarity_matrix}
\alias{compute_average_similarity_matrix}
\title{Compute Average Similarity Matrix}
\usage{
compute_average_similarity_matrix(percent_similarity_matrix)
}
\arguments{
\item{percent_similarity_matrix}{A matrix containing percent similarity values between sequences.
Rows and columns should be labeled with SequenceIDs.}
}
\value{
A matrix containing average similarity values between SampleIDs.
}
\description{
This function computes the average similarity matrix based on the percent similarity matrix.
}
\examples{
\dontrun{
# Assuming you have a percent_similarity_matrix
average_percent_similarity <- compute_average_similarity_matrix(percent_similarity_matrix)
print(average_percent_similarity)
}
}
