% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7heatmap_1.R
\name{generate_heatmaps}
\alias{generate_heatmaps}
\title{Generate Heatmaps}
\usage{
generate_heatmaps(
  average_percent_similarity,
  html_title = "Heatmap of Average Similarity",
  tiff_title = "Heatmap of Average Similarity",
  xlab = "SampleID",
  ylab = "SampleID",
  width_inch = 8,
  height_inch = 6,
  dpi = 300
)
}
\arguments{
\item{average_percent_similarity}{A matrix containing average similarity values between SampleIDs.
Rows and columns should be labeled with SampleIDs.}

\item{html_title}{Title for the interactive heatmap (HTML version).}

\item{tiff_title}{Title for the static heatmap (TIFF version).}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{width_inch}{Width of the heatmap in inches.}

\item{height_inch}{Height of the heatmap in inches.}

\item{dpi}{Dots per inch for the static heatmap.}
}
\value{
A list containing the file names of the generated heatmaps.
}
\description{
This function generates interactive and static heatmaps based on the average similarity matrix.
}
\examples{
\dontrun{
# Example usage:
output_directory <- "output_directory/new"
width_inch <- 8
height_inch <- 6
dpi <- 300

heatmap_files <- generate_heatmaps(average_percent_similarity)

# Save the interactive heatmap as an HTML file ####
html <- file.path(output_directory, "5. heatmap.html")
htmlwidgets should be installed and loaded
htmlwidgets::saveWidget(heatmap_files$html, file = html)

# save the TIFFE images ####

# heatmap_tiff_file1
tiff1 <- file.path(output_directory, "5. heatmap1.tiff")
tiff(tiff1, width = width_inch, height = height_inch, units = "in", res = dpi)
heatmap(as.matrix(average_percent_similarity), main = "Heatmap of Average Similarity")

# Close the TIFF device
dev.off()

# heatmap_tiff_file2
tiff2 <- file.path(output_directory, "5. heatmap2.tiff")

# Open the TIFF device
tiff(tiff2, width = width_inch, height = height_inch, units = "in", res = dpi)
print(heatmap_files$tiff2)

# Close the TIFF device
dev.off()

# heatmap_tiff_file3
tiff3 <- file.path(output_directory, "5. heatmap3.tiff")

# Open the TIFF device and create the heatmap.2 with hierarchical clustering dendrogram
# gplots should be installed and loaded
gplots::heatmap.2(as.matrix(average_percent_similarity),
          dendrogram = "row",
          Colv = "Rowv",
          scale = "row",
          main = "Heatmap of Average Similarity")

# Close the TIFF device
dev.off()
}
}
