% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8clustering_average_similarity.R
\name{clustering_average_similarity}
\alias{clustering_average_similarity}
\title{Perform Hierarchical Clustering for average similarity matrix}
\usage{
clustering_average_similarity(
  similarity_matrix,
  num_clusters_option = 10,
  cut_height_option = 0.2
)
}
\arguments{
\item{similarity_matrix}{A matrix containing similarity values between SampleIDs.}

\item{num_clusters_option}{The desired number of clusters. Default is 10.}

\item{cut_height_option}{The height at which the dendrogram should be cut to obtain clusters. Default is 0.2.}
}
\value{
A list containing the dendrogram, clustered data, sample percentage within clusters, and total percentage for each cluster.
}
\description{
This function performs hierarchical clustering based on a similarity matrix and provides additional information
such as a colored dendrogram, clustered data, sample percentage within clusters, and total percentage for each cluster.
}
\examples{
\dontrun{
# clustering_result <- clustering_average_similarity(average_percent_similarity)
# dend_colored <- clustering_result$dendrogram
# clustered_data <- clustering_result$clustered_data
# Cluster_SampleID_Percentage <- clustering_result$Cluster_SampleID_Percentage
# Cluster_TotalPercentage <- clustering_result$Cluster_TotalPercentage

output_directory <- "output_directory/new"
width_inch <- 8
height_inch <- 6
dpi <- 300

clustering_result <- clustering_average_similarity(average_percent_similarity)

# Extract the dendrogram and clustered data
dend_colored <- clustering_result$dendrogram
clustered_data <- clustering_result$clustered_data
Cluster_SampleID_Percentage <- clustering_result$Cluster_SampleID_Percentage
Cluster_TotalPercentage <- clustering_result$Cluster_TotalPercentage

tiff_file <- file.path(output_directory, "6. hierarchical_clustering_dendrogram_colored.tiff")

# Save the dendrogram as a TIFF image
tiff(tiff_file, width = width_inch, height = height_inch, units = "in", res = dpi)
plot(dend_colored, main = "Colored Hierarchical Clustering Dendrogram")
dev.off()

# Save the clustered data frame to a CSV file
write.csv(clustered_data, "output_directory/new/7. clustered_data.csv", row.names = FALSE)
}
}
