% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9bubble_plot1.R
\name{bubble_plot_count}
\alias{bubble_plot_count}
\title{Bubble Plot Count Function}
\usage{
bubble_plot_count(
  clustered_data,
  title,
  x_label,
  y_label,
  size_label,
  color_label
)
}
\arguments{
\item{clustered_data}{A data frame containing clustered data.}

\item{title}{The title of the plot.}

\item{x_label}{The label for the x-axis.}

\item{y_label}{The label for the y-axis.}

\item{size_label}{The label for the size variable.}

\item{color_label}{The label for the color variable.}
}
\value{
A ggplot object representing the bubble plot.
}
\description{
This function generates a bubble plot using ggplot2.
}
\examples{
\dontrun{
# Example usage with clustered_data
clustered_data <- clustered_data # Load or generate your clustered data
bubble_plot_count <- bubble_plot_count(clustered_data = clustered_data,
                                      title = "Bubble Plot of Clusters",
                                      x_label = "Clusters",
                                      y_label = "Sample ID",
                                      size_label = "Count",
                                      color_label = "Sample ID")

# Save the bubble plot as a TIFF image
output_directory <- "output_directory/new"
width_inch <- 8
height_inch <- 6
dpi <- 300

# Set the file name for the TIFF image
tiff_file <- file.path(output_directory, "bubble_plot_count.tiff")

# Open the TIFF device
tiff(tiff_file, width = width_inch, height = height_inch, units = "in", res = dpi)

# Print and save the bubble plot
print(bubble_plot_count)

# Close the TIFF device
dev.off()
}
}
