% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9bubble_plot2.R
\name{bubble_plot_percentage}
\alias{bubble_plot_percentage}
\title{Bubble Plot Percentage Function}
\usage{
bubble_plot_percentage(
  Cluster_SampleID_Percentage,
  title,
  x_label,
  y_label,
  size_label,
  color_label
)
}
\arguments{
\item{Cluster_SampleID_Percentage}{A data frame containing cluster, sample ID, and percentage data.}

\item{title}{The title of the plot.}

\item{x_label}{The label for the x-axis.}

\item{y_label}{The label for the y-axis.}

\item{size_label}{The label for the size variable.}

\item{color_label}{The label for the color variable.}
}
\value{
A ggplot object representing the bubble plot.
}
\description{
This function generates a bubble plot using ggplot2 based on percentage data.
}
\examples{
\dontrun{
# Example usage with Cluster_SampleID_Percentage
Cluster_SampleID_Percentage <- Cluster_SampleID_Percentage
bubble_plot_percentage <- bubble_plot_percentage(Cluster_SampleID_Percentage,
                                                title = "Bubble Plot",
                                                x_label = "Clusters",
                                                y_label = "Sample ID",
                                                size_label = "Percentage",
                                                color_label = "Sample ID")

# Save the bubble plot as a TIFF image
output_directory <- "output_directory/new"
width_inch <- 8
height_inch <- 6
dpi <- 300

# Set the file name for the TIFF image
tiff_file <- file.path(output_directory, "bubble_plot_percentage.tiff")

# Open the TIFF device
tiff(tiff_file, width = width_inch, height = height_inch, units = "in", res = dpi)

# Print and save the bubble plot
print(bubble_plot_percentage)

# Close the TIFF device
dev.off()
}
}
