% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10tree_percent_similarity.R
\name{tree_percent_similarity}
\alias{tree_percent_similarity}
\title{Generate Phylogenetic Tree and Color Palette for Percent Similarity Matrix}
\usage{
tree_percent_similarity(percent_similarity_matrix)
}
\arguments{
\item{percent_similarity_matrix}{A matrix containing pairwise percent similarities between samples.}
}
\value{
A list containing the phylogenetic tree and a color palette.
}
\description{
This function generates a Neighbor-Joining phylogenetic tree and a color palette based on the percent similarity matrix.
}
\examples{
\dontrun{
# Example usage with percent_similarity_matrix
result <- tree_percent_similarity(percent_similarity_matrix)
tree <- result$tree
color_palette <- result$color_palette

# Save the phylogenetic tree as a Newick file
write.tree(tree, file = "output_directory/new/phylogenetic_tree_nj.nwk")

# Save the phylogenetic tree as a TIFF image
output_directory <- "output_directory/new"
width_inch <- 8
height_inch <- 6
dpi <- 300

# Set the file name for the TIFF image
tiff_file <- file.path(output_directory, "phylogenetic_tree.tiff")

# Open the TIFF device
tiff(tiff_file, width = width_inch, height = height_inch, units = "in", res = dpi)

# Plot the phylogenetic tree vertically
plot(tree, main = "Neighbor-Joining Tree", cex = 1, direction = "downward")

# Close the TIFF device
dev.off()

# Set the file name for the TIFF image with rainbow-colored branches
tiff_file_rainbow <- file.path(output_directory, "phylogenetic_tree_rainbow.tiff")

# Open the TIFF device
tiff(tiff_file_rainbow, width = width_inch, height = height_inch, units = "in", res = dpi)

# Plot the phylogenetic tree vertically with rainbow-colored branches
plot(tree, main = "NJ Tree", cex = 1, direction = "downward", tip.color = color_palette)

# Close the TIFF device
dev.off()
}
}
