\name{plot.gama}
\alias{plot.gama}
\title{Plots results of a Gama clustering.}
\description{
  This function takes a gama object and plots the partitions found by the algorithm. The partitions will be coloured accordingly the distribution of the clusters.
}
\usage{
\method{plot}{gama}(x, view.method = "pca", \dots)
}
\arguments{
  \item{x}{an object of type 'gama' generated by an appropriate gama call to gama clustering.}
  \item{view.method}{the representation used to plot the partitions. If 'pca' is used, the function will perform a Principal Component Analysis over the dataset and plot the two main components. If 'total.sum' is used, the function will sum all dimensions for each observation and plot this sum, highlighting its partitions. If ommited, the default value 'pca' will used.}
  \item{\dots}{other arguments that user may pass to the function.}

}
\references{
Mardia, K. V., J. T. Kent, and J. M. Bibby (1979) Multivariate Analysis, London: Academic Press.
}
\seealso{
  \code{\link{gama}, \link{print.gama}}.
}

\examples{

# loads data about CPU execution metrics of a distributed
# version of Alternating Least Squares (ALS) algorithm
data(cpu.als)

# call the gama clustering algorithm
gamaObj <- gama(cpu.als, k = 2)


# call the plot.gama function with "Principal Component Analysis" view method
# note: plot.gama uses generic function concept, which allows a call to plot, only.
plot(gamaObj, view.method = "pca")

# call the plot.gama function with "Total Sum" of dimensions view method
plot(gamaObj, view.method = "total.sum")

}

\keyword{file}
