% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_availability}
\alias{plot_availability}
\title{Plot data richness}
\usage{
plot_availability(
  SUMMARY,
  rank = c("highest", "lowest", "A-Z", "Z-A", "input"),
  abbreviate = TRUE,
  theme_fn = ggplot2::theme_minimal,
  colours = c(Assembly = "#E3F9F2", SRA = "#A8E6CF", BioSample = "#56C5A8")
)
}
\arguments{
\item{SUMMARY}{Tibble returned by \code{\link[=summarise_availability]{summarise_availability()}}.}

\item{rank}{Character; species ordering method.}

\item{abbreviate}{Logical; abbreviate species names.}

\item{theme_fn}{ggplot2 theme function.}

\item{colours}{Named character vector of fill colours.}
}
\value{
A ggplot object showing stacked bar charts of data richness components.
}
\description{
Plot data richness
}
