% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_sra_availability}
\alias{plot_sra_availability}
\title{Plot proportional SRA experimental modality composition}
\usage{
plot_sra_availability(
  SRA,
  species = NULL,
  rank = c("highest", "lowest", "A-Z", "Z-A", "input"),
  abbreviate = TRUE,
  theme_fn = ggplot2::theme_minimal,
  colours = c(genomic = "#E3F9F2", transcriptomic = "#A8E6CF", epigenomic = "#56C5A8",
    chromatin = "#2FA083", other = "#166A55", unknown = "#BDBDBD")
)
}
\arguments{
\item{SRA}{Tibble returned by \code{\link[=summarise_sra_availability]{summarise_sra_availability()}}.}

\item{species}{Optional character vector of species to include.}

\item{rank}{Character; species ordering method.}

\item{abbreviate}{Logical; abbreviate species names.}

\item{theme_fn}{ggplot2 theme function.}

\item{colours}{Named character vector of class colours.}
}
\value{
A ggplot object showing proportional SRA modality composition across species.
}
\description{
Plot proportional SRA experimental modality composition
}
