% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{print.gdt_tbl}
\alias{print.gdt_tbl}
\title{Print GAMA summary tibble}
\usage{
\method{print}{gdt_tbl}(x, ...)
}
\arguments{
\item{x}{An object of class 'gdt_tbl'.}

\item{...}{Further arguments passed to the next print method.}
}
\value{
The input object, invisibly.
}
\description{
Print GAMA summary tibble
}
