% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Availability.R
\name{summarise_availability}
\alias{summarise_availability}
\title{Summarise accession counts and compute data richness scores}
\usage{
summarise_availability(results)
}
\arguments{
\item{results}{List returned by \code{\link[=query_species]{query_species()}}.}
}
\value{
A tibble of per-species accession counts and composite data richness scores (class 'gdt_tbl').
}
\description{
Summarise accession counts and compute data richness scores
}
